requirements.txt
flask>=2.3.3
werkzeug>=2.3.7
flask-cors>=4.0.0
python-dotenv>=1.0.0
openai==1.12.0
azure-cognitiveservices-speech==1.34.0
pydub==0.25.1
requests>=2.31.0
azure-ai-textanalytics>=5.3.0 
